requirements.txt
ultralytics 
supervision
streamlit
gdown
cvzone
pyyaml
opencv-python